fix: iteration
