+++ apps/api-languages/db/seed.ts
@@ -1,8 +1,8 @@
 import { aql } from 'arangojs'
 import { float } from 'aws-sdk/clients/lightsail'
-import { getSeoSlug } from '@core/api-seed'
 import { isEmpty } from 'lodash'
 import fetch from 'node-fetch'
+import slugify from 'slugify'
 import { ArangoDB } from './db'
 
 interface Language {
@@ -171,6 +171,25 @@ async function getCountries(language = 'en'): Promise<MediaCountry[]> {
 }
 
 const usedTitles: string[] = []
+
+export function getIteration(slug: string, collection: string[]): string {
+  const exists = collection.find((t) => t === slug)
+  if (exists != null && slug !== '') {
+    const regex = slug.match(/^(.*?)-(\d+)$/)
+    const iteration = parseInt(regex?.[2] ?? '1') + 1
+    const title = regex?.[1] ?? slug
+    const value = `${title}-${iteration}`
+    return getIteration(value, collection)
+  }
+  return slug
+}
+
+export function getSeoSlug(title: string, collection: string[]): string {
+  const slug = slugify(title, { lower: true, remove: /[^a-zA-Z\d\s:]/g })
+  const newSlug = getIteration(slug, collection)
+  collection.push(newSlug)
+  return newSlug
+}
 
 function digestCountries(countries: MediaCountry[]): Country[] {
   console.log('countries:', '529')
