--- libs/api-seed/README.md
@@ -1,7 +1,0 @@
-# api-seed
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test api-seed` to execute the unit tests via [Jest](https://jestjs.io).
