--- libs/api-seed/project.json
@@ -1,23 +1,0 @@
-{
-  "root": "libs/api-seed",
-  "sourceRoot": "libs/api-seed/src",
-  "projectType": "library",
-  "targets": {
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["libs/api-seed/**/*.ts"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/api-seed"],
-      "options": {
-        "jestConfig": "libs/api-seed/jest.config.js",
-        "passWithNoTests": true
-      }
-    }
-  },
-  "tags": []
-}
