--- libs/api-seed/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/api-seed'
